+++ plugin.yml
@@ -1,0 +1,11 @@
+name: Wither
+main: com.mutinycraft.jigsaw.Wither.Wither
+version: 2.1
+author: Jigsaw
+description: Control things related to the Wither.
+website: http://dev.bukkit.org/server-mods/wither/
+
+commands:
+   wither:
+      description: Allows you to reload the plugin.
+      usage: /wither reload
